Moved CohortDto and related converters to appropriate package
