--- src/main/java/org/ohdsi/webapi/cohortcharacterization/dto/CohortDetailsDTO.java
@@ -1,8 +1,0 @@
-package org.ohdsi.webapi.cohortcharacterization.dto;
-
-import org.ohdsi.circe.cohortdefinition.CohortExpression;
-
-public class CohortDetailsDTO {
-    
-    
-}
